[ito] fix dag profiler for parallel reduction
An optimization in parallel_reduce makes the first thread be joined
before spawning the last thread (when the parent is not stolen), but
this case was not considered in the dag profiler. To fix this issue,
task group information is desired to be accessed at any time (other
than task_group_begin/end time), so now task group data have a pointer
to its parent (only within the same thread).